feat: native typo3 backend module (#35)

+++ Classes/Domain/Repository/SentMessageRepository.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Fab\Messenger\Domain\Repository;
 
 /*
@@ -11,24 +12,99 @@ namespace Fab\Messenger\Domain\Repository;
 use Fab\Messenger\Utility\Algorithms;
 use Fab\Vidi\Tca\Tca;
 use RuntimeException;
-use TYPO3\CMS\Core\Database\ConnectionPool;
-use TYPO3\CMS\Core\Database\Query\QueryBuilder;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 
-/**
- * A repository for handling sent message
- */
-class SentMessageRepository
+class SentMessageRepository extends AbstractContentRepository
 {
+    protected string $tableName = 'tx_messenger_domain_model_sentmessage';
 
-    /**
-     * @var string
-     */
-    protected $tableName = 'tx_messenger_domain_model_sentmessage';
+    protected QueryInterface $constraints;
 
-    /**
-     * @throws RuntimeException
-     */
+    public function findByUid(int $uid): array
+    {
+        $query = $this->getQueryBuilder();
+        $query
+            ->select('*')
+            ->from($this->tableName)
+            ->where(
+                $this->getQueryBuilder()
+                    ->expr()
+                    ->eq('uid', $this->getQueryBuilder()->expr()->literal($uid)),
+            );
+
+        $messages = $query->execute()->fetch();
+
+        return is_array($messages) ? $messages : [];
+    }
+
+    public function findByUuid(string $uuid): array
+    {
+        $query = $this->getQueryBuilder();
+        $query
+            ->select('*')
+            ->from($this->tableName)
+            ->where(
+                $this->getQueryBuilder()
+                    ->expr()
+                    ->eq('uuid', $this->getQueryBuilder()->expr()->literal($uuid)),
+            );
+
+        $messages = $query->execute()->fetch();
+
+        return is_array($messages) ? $messages : [];
+    }
+
+    public function findOlderThanDays(int $days): array
+    {
+        $time = time() - $days * 86400;
+        $query = $this->getQueryBuilder();
+        $query
+            ->select('*')
+            ->from($this->tableName)
+            ->where('crdate < ' . $time);
+
+        $messages = $query->execute()->fetchAll();
+        return is_array($messages) ? $messages : [];
+    }
+
+    public function removeOlderThanDays(int $days): int
+    {
+        $time = time() - $days * 86400;
+
+        $query = $this->getQueryBuilder();
+        $query->delete($this->tableName)->where('crdate < ' . $time);
+
+        return $query->execute();
+    }
+
+    public function findByDemand(array $demand = [], array $orderings = [], int $offset = 0, int $limit = 0): array
+    {
+        $queryBuilder = $this->getQueryBuilder();
+        $queryBuilder->select('*')->from($this->tableName);
+
+        $constraints = [];
+        foreach ($demand as $field => $value) {
+            $constraints[] = $queryBuilder
+                ->expr()
+                ->like(
+                    $field,
+                    $queryBuilder->createNamedParameter('%' . $queryBuilder->escapeLikeWildcards($value) . '%'),
+                );
+        }
+        if ($constraints) {
+            $queryBuilder->where($queryBuilder->expr()->orX(...$constraints));
+        }
+
+        # We handle the sorting
+        $queryBuilder->addOrderBy(key($orderings), current($orderings));
+
+        if ($limit > 0) {
+            $queryBuilder->setMaxResults($limit);
+        }
+
+        return $queryBuilder->execute()->fetchAllAssociative();
+    }
+
     public function add(array $message): int
     {
         $values = [];
@@ -49,8 +125,7 @@ class SentMessageRepository
         }
 
         $query = $this->getQueryBuilder();
-        $query->insert($this->tableName)
-            ->values($values);
+        $query->insert($this->tableName)->values($values);
 
         $result = $query->execute();
         if (!$result) {
@@ -59,73 +134,10 @@ class SentMessageRepository
         return $result;
     }
 
-    public function findByUid(int $uid): array
+    public function removeByUid(int $uid): int
     {
         $query = $this->getQueryBuilder();
-        $query->select('*')
-            ->from($this->tableName)
-            ->where(
-                $this->getQueryBuilder()->expr()->eq(
-                    'uid',
-                    $this->getQueryBuilder()->expr()->literal($uid)
-                )
-            );
-
-        $messages = $query->execute()->fetch();
-
-        return is_array($messages) ? $messages : [];
-    }
-
-    public function findByUuid(string $uuid): array
-    {
-        $query = $this->getQueryBuilder();
-        $query->select('*')
-            ->from($this->tableName)
-            ->where(
-                $this->getQueryBuilder()->expr()->eq(
-                    'uuid',
-                    $this->getQueryBuilder()->expr()->literal($uuid)
-                )
-            );
-
-        $messages = $query->execute()->fetch();
-
-        return is_array($messages) ? $messages : [];
-    }
-
-    public function findOlderThanDays(int $days): array
-    {
-        $time = time() - ($days * 86400);
-        $query = $this->getQueryBuilder();
-        $query->select('*')
-            ->from($this->tableName)
-            ->where(
-                'crdate < ' . $time
-            );
-
-        $messages = $query->execute()->fetchAll();
-        return is_array($messages) ? $messages : [];
-    }
-
-    public function removeOlderThanDays(int $days): int
-    {
-        $time = time() - ($days * 86400);
-
-        $query = $this->getQueryBuilder();
-        $query->delete($this->tableName)
-            ->where('crdate < ' . $time);
-
+        $query->delete($this->tableName)->where('uid = ' . $uid);
         return $query->execute();
     }
-
-    /**
-     * @return object|QueryBuilder
-     */
-    protected function getQueryBuilder(): QueryBuilder
-    {
-        /** @var ConnectionPool $connectionPool */
-        $connectionPool = GeneralUtility::makeInstance(ConnectionPool::class);
-        return $connectionPool->getQueryBuilderForTable($this->tableName);
-    }
-
 }
